Move py eval fn
