+++ docs/reference/python-integration.md
@@ -105,7 +105,7 @@ assert egraph.eval(evalled) == 3
 
 ### Simpler Eval
 
-Instead of using the above low level primitive for evaling, there is a higher level wrapper function, `egraph.eval_fn`.
+Instead of using the above low level primitive for evaling, there is a higher level wrapper function, `eval_fn`.
 
 It takes in a Python function and converts it to a function of PyObjects, by using `py_eval`
 under the hood.
@@ -116,7 +116,7 @@ The above code code be re-written like this:
 def my_add(a, b):
     return a + b
 
-evalled = egraph.eval_fn(lambda a: my_add(a, 2))(1)
+evalled = eval_fn(lambda a: my_add(a, 2))(1)
 assert egraph.eval(evalled) == 3
 ```
